MLOPS/cars/requirements_airflow.txt
apache-airflow==3.1.2
pandas==2.2.3
requests==2.32.3
scikit-learn==1.6.1
